+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Revision history for http-interchange
 
+## 0.3.0.0 -- 2023-??-??
+
+* Add `Http.Headers.fromList`. In examples, this makes it easier
+  to create HTTP requests.
+* Add `Eq` and `Show` instances for `Bodied`.
+* Get rid of `versionMajor` and `versionMinor`.
+
 ## 0.2.0.0 -- 2023-08-14
 
 * Redo module structure